+++ lib/presentation/components/quiz/quiz_controller.dart
@@ -1,6 +1,7 @@
 // controllers/quiz_controller.dart
 import 'dart:async';
 import 'package:accredit/domain/services/api_asodya_manager.dart';
+import 'package:accredit/domain/services/api_certification_manager.dart';
 import 'package:flutter/foundation.dart';
 import 'package:accredit/core/utils/my_logs.dart';
 import 'package:accredit/domain/models/quiz.dart';
@@ -97,10 +98,16 @@ class QuizController extends ChangeNotifier {
 
     try {
       // Example side-effect call
-      final resp = await ApiAsodyaManager().updateUserInfo(formData.fullName, formData.phoneE164);
-      debug('updateUserInfo => ${resp.statusCode}');
+      
 
-      // TODO: await your quiz submit here with `payload` (+ formData, timeSpent, etc.)
+      final resp = await ApiAsodyaManager().updateUserInfo(formData.fullName, formData.phoneE164);
+      
+
+      // set some duration to simulate network delay
+
+      await Future.delayed(const Duration(seconds: 3));
+
+      final _ = await CertificationManager().submitQuiz(payload, timeSpent, formData);
 
       _finishCompleter!.complete(result);
       return result;