+++ lib/presentation/widgets/attachment/mobile_attachment.dart
@@ -76,7 +76,7 @@ class _MobileAttachmentState extends BaseAttachmentState<MobileAttachment> {
                                 body,
                                 const SizedBox(height: 10),
                                 WantText(onRequest: (url) async {
-                                  final manager = CardItemsManager();
+                                  final manager = CertificationManager();
                                   final response =
                                       await manager.requestNewCards(url);
                                       if (response.statusCode == 201) {}