+++ CHANGELOG.md
@@ -64,6 +64,7 @@ All notable changes to this project will be documented in this file.
 - *(loading)* Adjust loading design for others
 - *(topics)* Added the custom loading for topics cards too
 - *(pdf)* Added the custom loading for pdf too
+- *(quiz)* Adjusted headers for quiz submit
 
 ### 🐛 Bug Fixes